+++ api.planx.uk/modules/send/createSendEvents/controller.ts
@@ -1,21 +1,11 @@
-import { NextFunction, Request, Response } from "express";
 import {
-  ScheduledEventResponse,
+  CombinedResponse,
   createScheduledEvent,
 } from "../../../lib/hasura/metadata";
-
-interface CombinedResponse {
-  bops?: ScheduledEventResponse;
-  uniform?: ScheduledEventResponse;
-  email?: ScheduledEventResponse;
-}
+import { CreateSendEventsController } from "./types";
 
 // Create "One-off Scheduled Events" in Hasura from Send component for selected destinations
-const createSendEvents = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-): Promise<NextFunction | Response | void> => {
+const createSendEvents: CreateSendEventsController = async (req, res, next) => {
   try {
     const now = new Date();
     const combinedResponse: CombinedResponse = {};
